Merged old and new Debian "default" files
On recent versions of Ubuntu "start" variable in /etc/default/monit did not work
and monit was never enabled on boot. Merged the old and new style default files
together to produce an "universal" version. This avoids having to detect the
exact Debian/Ubuntu version to copy over a correct file.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>