Change cycle count and remove timeouts from directory-size.monit.erb
Monit 5.4 and 5.6 have a different syntax for defining "check program" timeouts.
As the default timeout value of 5 minutes is reasonable, the explicit timeout
value was removed to easily support both monit versions. In addition the cycle
count was increased to 5, so that periodic directory size fluctuations don't
trigger false positives as easily.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>