Start using conf-enabled as the fragment directory on latest Debian/Ubuntu
Both Debian 9 and Ubuntu 16.04 have moved to using /etc/monit/conf-enabled as
the fragment directory, and that directory is empty by default. So it can be
safely reused for Puppet-managed monit fragments. Earlier (LTS) versions used
/etc/monit/conf.d. The one exception was Debian 8 which uses
/etc/monit/monitrc.d that was by default populated with lots of files, so this
module continues to use conf.d on that platform. This change hopefully reduces
the surprise factor of Puppet-managed monit configurations, moves on with times
and improves compatibility with the Puppet-Finland/ansible-monit role.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>